# api/openapi.yaml
openapi: 3.1.0
info:
  title: capp-backend API
  description: |
    Multi-cluster API gateway for managing Capp (container-app) resources on
    Kubernetes / OpenShift clusters.

    ## Authentication

    The backend supports five auth modes configured server-side:

    - **passthrough** — supply a raw Kubernetes `Bearer <token>` on every request.
    - **jwt** — call `/api/v1/auth/login` with `cluster` + `token` to exchange a
      cluster token for a short-lived JWT; refresh with `/api/v1/auth/refresh`.
    - **static** — use a fixed API key configured on the server (dev / CI only).
    - **dex** — call `/api/v1/auth/login` with `username` + `password`; Dex handles OIDC.
    - **openshift** — use `/api/v1/auth/openshift/authorize` to get an OAuth URL,
      then POST the code to `/api/v1/auth/openshift/callback`.

    Use `GET /api/v1/auth/mode` to discover which mode the server is running.
    All resource endpoints require the `Authorization: Bearer <token>` header.
  version: "0.1.0"
  contact:
    name: dana-team
    url: https://github.com/dana-team/capp-backend
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        In **passthrough** mode supply the raw Kubernetes ServiceAccount token.
        In **jwt** / **dex** / **openshift** modes supply the JWT returned by the
        login or callback endpoint.
        In **static** mode supply one of the configured API keys.
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: resource "my-capp" not found
            status:
              type: integer
              example: 404

    # ── Auth ─────────────────────────────────────────────────────────────────────
...
    CappSize:
      type: string
      enum: [small, medium, large]
      description: |
        High-level sizing abstraction. Maps to specific CPU/memory requests and
        limits configured on the server. Mutually exclusive with setting raw
        resource quantities directly.
      example: small
    KeySelector:
      type: object
      required: [name, key]
      properties:
        name:
          type: string
          description: Name of the Secret or ConfigMap.
          example: my-secret
        key:
          type: string
          description: Key within the Secret or ConfigMap.
          example: password
    EnvVarSource:
      type: object
      description: |
        Selects the value of an env var from a Secret or ConfigMap key.
        Exactly one of `secretKeyRef` or `configMapKeyRef` must be set.
      properties:
        secretKeyRef:
          $ref: '#/components/schemas/KeySelector'
        configMapKeyRef:
          $ref: '#/components/schemas/KeySelector'
    EnvVar:
      type: object
      required: [name]
      description: |
        `value` and `valueFrom` are mutually exclusive. Set `value` for a literal
        string or `valueFrom` to reference a Secret/ConfigMap key. Setting both
        is invalid.
      properties:
        name:
          type: string
          example: LOG_LEVEL
        value:
          type: string
          example: info
        valueFrom:
          $ref: '#/components/schemas/EnvVarSource'
    VolumeMount:
      type: object
      required: [name, mountPath]
      properties:
        name:
          type: string
          example: config-vol
        mountPath:
          type: string
          example: /etc/config
    RouteSpec:
      type: object
      properties:
        hostname:
          type: string
          example: my-app.example.com
        tlsEnabled:
          type: boolean
          example: true
        routeTimeoutSeconds:
          type: integer
          format: int64
          description: Maximum request duration in seconds.
          example: 30
    LogSpec:
      type: object
      properties:
        type:
          type: string
          enum: [elastic, elastic-datastream]
          description: Log destination type. When "elastic-datastream", the index field is ignored.
          example: elastic
        host:
          type: string
          example: https://elasticsearch.example.com:9200
        index:
          type: string
          description: Index name to write events to. Ignored when type is "elastic-datastream".
          example: capp-logs
        user:
          type: string
          example: elastic-user
        passwordSecret:
          type: string
          description: Name of the Kubernetes Secret holding the Elasticsearch password.
          example: elastic-password-secret
    NFSVolume:
      type: object
      required: [name, server, path, capacity]
      properties:
        name:
          type: string
          example: nfs-vol
        server:
          type: string
          example: nfs.example.com
        path:
          type: string
          example: /exports/data
        capacity:
          type: string
          description: Storage capacity, e.g. "10Gi".
          example: "10Gi"
    SecretVolume:
      type: object
      required: [name, secretName, mountPath]
      properties:
        name:
          type: string
          example: my-secret-vol
        secretName:
          type: string
          description: Name of the Kubernetes Secret to mount.
          example: my-secret
        mountPath:
          type: string
          example: /etc/secrets
    ConfigMapVolume:
      type: object
      required: [name, configMapName, mountPath]
      properties:
        name:
          type: string
          example: my-cm-vol
        configMapName:
          type: string
          description: Name of the Kubernetes ConfigMap to mount.
          example: my-config
        mountPath:
          type: string
          example: /etc/config
    ScaleSpec:
      type: object
      properties:
        metric:
          type: string
          enum: [rps, concurrency, cpu, memory]
          example: concurrency
        minReplicas:
          type: integer
          minimum: 0
          example: 1
        scaleDelaySeconds:
          type: integer
          minimum: 0
          example: 30
    CappRequest:
      type: object
      required: [name, namespace, image]
      properties:
        name:
          type: string
          example: my-app
        namespace:
          type: string
          example: default
        scaleSpec:
          $ref: '#/components/schemas/ScaleSpec'
        state:
          type: string
          enum: [enabled, disabled]
          example: enabled
        image:
          type: string
          example: ghcr.io/example/my-app:v1.2.3
        containerName:
          type: string
          example: my-app
        size:
          $ref: '#/components/schemas/CappSize'
        env:
          type: array
          items:
            $ref: '#/components/schemas/EnvVar'
        volumeMounts:
          type: array
          items:
            $ref: '#/components/schemas/VolumeMount'
        routeSpec:
          $ref: '#/components/schemas/RouteSpec'
        logSpec:
          $ref: '#/components/schemas/LogSpec'
        nfsVolumes:
          type: array
          items:
            $ref: '#/components/schemas/NFSVolume'
        secretVolumes:
          type: array
          items:
            $ref: '#/components/schemas/SecretVolume'
        configMapVolumes:
          type: array
          items:
            $ref: '#/components/schemas/ConfigMapVolume'
        eventSourcesSpec:
          $ref: '#/components/schemas/EventSourcesSpec'
    ConditionResponse:
      type: object
      properties:
        source:
          type: string
          description: |
            Sub-system that produced this condition.
            One of: "capp", "knative", "logging", "route.domainmapping".
          example: knative
        type:
          type: string
          example: Ready
        status:
          type: string
          enum: ["True", "False", Unknown]
          example: "True"
        lastTransitionTime:
          type: string
          format: date-time
        reason:
          type: string
          example: RevisionReady
        message:
          type: string
          example: Revision is ready.
    PingSourceConfig:
      type: object
      required: [schedule]
      properties:
        schedule:
          type: string
          description: Cron expression defining when the PingSource fires.
          example: "*/5 * * * *"
        data:
          type: string
          description: JSON payload sent on each trigger. Must be valid JSON.
          example: '{"msg":"hello"}'
    KafkaSourceConfig:
      type: object
      required: [bootstrapServers, topics, secretRef]
      properties:
        bootstrapServers:
          type: array
          minItems: 1
          items:
            type: string
          example: ["broker1:9092", "broker2:9092"]
        topics:
          type: array
          minItems: 1
          items:
            type: string
          example: ["topic-a"]
        consumerGroup:
          type: string
          description: Kafka consumer group ID. Optional; operator assigns one when unset.
          example: my-group
        consumers:
          type: integer
          minimum: 1
          description: Number of parallel consumers for this source.
          example: 2
        secretRef:
          type: string
          description: Name of the Secret in the Capp namespace with Kafka credentials.
          example: kafka-creds
    SourceConfig:
      type: object
      required: [name]
      description: |
        A single Knative Eventing source attached to the Capp.
        Exactly one of pingSourceConfiguration or kafkaSourceConfiguration must be set.
      properties:
        name:
          type: string
          description: Unique logical name for this source within the Capp.
          example: ping-src
        uri:
          type: string
          description: Relative path appended to the Capp Knative Service address (e.g. "/events").
          example: /events
        pingSourceConfiguration:
          $ref: '#/components/schemas/PingSourceConfig'
        kafkaSourceConfiguration:
          $ref: '#/components/schemas/KafkaSourceConfig'
    EventSourcesSpec:
      type: object
      properties:
        sources:
          type: array
          items:
            $ref: '#/components/schemas/SourceConfig'
    EventSourceStatusResponse:
      type: object
      properties:
        name:
          type: string
          description: Name of the underlying event source resource.
          example: ping-src
        type:
          type: string
          description: Condition type from the underlying source resource.
          example: Ready
        status:
          type: string
          enum: ["True", "False", Unknown]
          example: "True"
        reason:
          type: string
          example: EventSourceReady
        message:
          type: string
          example: PingSource is ready.
    EventingStatusResponse:
      type: object
      properties:
        eventSources:
          type: array
          items:
            $ref: '#/components/schemas/EventSourceStatusResponse'
    StateStatusResponse:
      type: object
      properties:
        state:
          type: string
          enum: [enabled, disabled]
          example: enabled
        lastChange:
          type: string
          format: date-time
          description: RFC3339 timestamp of the last state transition.
    CappStatusResponse:
      type: object
      properties:
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/ConditionResponse'
        eventingStatus:
          $ref: '#/components/schemas/EventingStatusResponse'
        stateStatus:
          $ref: '#/components/schemas/StateStatusResponse'
    CappResponse:
      type: object
      properties:
        name:
          type: string
          example: my-app
        namespace:
          type: string
          example: default
        createdAt:
          type: string
          format: date-time
        uid:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        resourceVersion:
          type: string
          example: "12345"
        labels:
          type: object
          additionalProperties:
            type: string
        annotations:
          type: object
          additionalProperties:
            type: string
          description: kubectl.kubernetes.io/* annotations are filtered out.
        scaleSpec:
          $ref: '#/components/schemas/ScaleSpec'
        state:
          type: string
          enum: [enabled, disabled]
          example: enabled
        image:
          type: string
          example: ghcr.io/example/my-app:v1.2.3
        containerName:
          type: string
          example: my-app
        size:
          $ref: '#/components/schemas/CappSize'
        env:
          type: array
          items:
            $ref: '#/components/schemas/EnvVar'
        volumeMounts:
          type: array
          items:
            $ref: '#/components/schemas/VolumeMount'
        routeSpec:
          $ref: '#/components/schemas/RouteSpec'
        logSpec:
          $ref: '#/components/schemas/LogSpec'
        nfsVolumes:
          type: array
          items:
            $ref: '#/components/schemas/NFSVolume'
        secretVolumes:
          type: array
          items:
            $ref: '#/components/schemas/SecretVolume'
        configMapVolumes:
          type: array
          items:
            $ref: '#/components/schemas/ConfigMapVolume'
        eventSourcesSpec:
          $ref: '#/components/schemas/EventSourcesSpec'
        status:
          $ref: '#/components/schemas/CappStatusResponse'
    CappList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CappResponse'
        total:
          type: integer
          example: 3
    SyncResponse:
      type: object
      properties:
        commitSha:
          type: string
          description: Git commit SHA of the pushed values file.
          example: "a1b2c3d4e5f6"
        path:
          type: string
          description: Relative path of the values file in the GitOps repository.
          example: "namespaces/default/my-app/values.yaml"
  parameters:
    clusterParam:
      name: cluster
      in: path
      required: true
      schema:
        type: string
      description: Cluster name as configured on the backend.
      example: production
    namespaceParam:
      name: namespace
      in: path
      required: true
      schema:
        type: string
      description: Kubernetes namespace.
      example: default
    objectNameParam:
      name: name
      in: path
      required: true
      schema:
        type: string
      description: Object resource name.
      example: my-app
  responses:
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The caller does not have permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: A resource with this name already exists.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - bearerAuth: []
paths:
...
  /api/v1/clusters/{cluster}/capps:
    get:
      tags: [capps]
      summary: List all Capps across all namespaces
      description: |
        Returns Capps from every namespace the authenticated user can access.
        Useful for cluster-wide dashboards.
      parameters:
        - $ref: '#/components/parameters/clusterParam'
      responses:
        "200":
          description: Capp list (all namespaces).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CappList'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "503":
          description: Cluster is currently unhealthy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/clusters/{cluster}/namespaces/{namespace}/capps:
    get:
      tags: [capps]
      summary: List Capps in a namespace
      parameters:
        - $ref: '#/components/parameters/clusterParam'
        - $ref: '#/components/parameters/namespaceParam'
      responses:
        "200":
          description: Capp list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CappList'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "503":
          description: Cluster is currently unhealthy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [capps]
      summary: Create a Capp
      parameters:
        - $ref: '#/components/parameters/clusterParam'
        - $ref: '#/components/parameters/namespaceParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CappRequest'
      responses:
        "201":
          description: Capp created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CappResponse'
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
  /api/v1/clusters/{cluster}/namespaces/{namespace}/capps/{name}:
    get:
      tags: [capps]
      summary: Get a Capp
      parameters:
        - $ref: '#/components/parameters/clusterParam'
        - $ref: '#/components/parameters/namespaceParam'
        - $ref: '#/components/parameters/objectNameParam'
      responses:
        "200":
          description: Capp details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CappResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

    put:
      tags: [capps]
      summary: Update a Capp
      description: |
        Replaces the Capp spec. The backend fetches the current `resourceVersion`
        before applying the update to satisfy Kubernetes optimistic concurrency control.
      parameters:
        - $ref: '#/components/parameters/clusterParam'
        - $ref: '#/components/parameters/namespaceParam'
        - $ref: '#/components/parameters/objectNameParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CappRequest'
      responses:
        "200":
          description: Capp updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CappResponse'
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [capps]
      summary: Delete a Capp
      parameters:
        - $ref: '#/components/parameters/clusterParam'
        - $ref: '#/components/parameters/namespaceParam'
        - $ref: '#/components/parameters/objectNameParam'
      responses:
        "204":
          description: Capp deleted.
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /api/v1/clusters/{cluster}/namespaces/{namespace}/capps/{name}/sync:
    post:
      tags: [capps]
      summary: Sync a Capp to the GitOps repository
      description: |
        Generates a Helm values file from the live Capp spec and pushes it to the
        configured GitOps repository. Also patches the `rcs.dana.io/backup-to-git=true`
        label onto the Capp (idempotent on re-calls).

        Returns `501 Not Implemented` when the backend is not configured with a
        GitOps integration (`gitops.enabled: false`).
      parameters:
        - $ref: '#/components/parameters/clusterParam'
        - $ref: '#/components/parameters/namespaceParam'
        - $ref: '#/components/parameters/objectNameParam'
      responses:
        "200":
          description: Capp values pushed to GitOps repository.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "501":
          description: GitOps integration is not enabled on this backend.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          $ref: '#/components/responses/InternalError'
        "503":
          description: Cluster is currently unhealthy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Observability ──────────────────────────────────────────────────────────────
...